refactor: extract single model files to out of folders
