+++ src/data/protocols/exam-repository/get-exam-repository.ts
@@ -1,4 +1,4 @@
-import { ExamModel } from '../../../domain/models/exam/exam-model'
+import { ExamModel } from '../../../domain/models/exam-model'
 
 export interface GetExamRepository {
   get: (id: string) => Promise<ExamModel>
