+++ src/data/protocols/exam-repository/list-exams-repository.ts
@@ -1,4 +1,4 @@
-import { ExamModel } from '../../../domain/models/exam/exam-model'
+import { ExamModel } from '../../../domain/models/exam-model'
 
 export interface ListExamsRepository {
   list: () => Promise<ExamModel[]>
