+++ src/data/protocols/exam-repository/update-exam-repository.ts
@@ -1,4 +1,4 @@
-import { ExamModel } from '../../../domain/models/exam/exam-model'
+import { ExamModel } from '../../../domain/models/exam-model'
 import { UpdateExamModel } from '../../../domain/usecases/exam/update-exam'
 
 export interface UpdateExamRepository {
