+++ src/data/protocols/question-repository/add-question-repository.ts
@@ -1,4 +1,4 @@
-import { QuestionModel } from '../../../domain/models/question/question-model'
+import { QuestionModel } from '../../../domain/models/question-model'
 import { AddQuestionModel } from '../../../domain/usecases/question/add-question'
 
 export interface AddQuestionRepository {
