+++ src/data/protocols/question-repository/get-question-repository.ts
@@ -1,4 +1,4 @@
-import { QuestionModel } from '../../../domain/models/question/question-model'
+import { QuestionModel } from '../../../domain/models/question-model'
 import { GetQuestionModel } from '../../../domain/usecases/question/get-question'
 
 export interface GetQuestionRepository {
