+++ src/data/protocols/question-repository/list-questions-repository.ts
@@ -1,4 +1,4 @@
-import { QuestionModel } from '../../../domain/models/question/question-model'
+import { QuestionModel } from '../../../domain/models/question-model'
 import { ListQuestionsModel } from '../../../domain/usecases/question'
 
 export interface ListQuestionsRepository {
