+++ src/data/usecases/exam/add-exam/db-add-exam-protocols.ts
@@ -1,3 +1,3 @@
-export * from '../../../../domain/models/exam/exam-model'
+export * from '../../../../domain/models/exam-model'
 export * from '../../../../domain/usecases/exam/add-exam'
 export * from '../../../protocols/exam-repository/add-exam-repository'
