+++ src/data/usecases/exam/add-exam/db-add-exam.spec.ts
@@ -1,4 +1,4 @@
-import { ExamModel } from '../../../../domain/models/exam/exam-model'
+import { ExamModel } from '../../../../domain/models/exam-model'
 import { AddExamModel } from '../../../../domain/usecases/exam/add-exam'
 import { AddExamRepository } from '../../../protocols/exam-repository/add-exam-repository'
 import { DbAddExam } from './db-add-exam'
