+++ src/data/usecases/exam/delete-exam/db-delete-exam-protocols.ts
@@ -1,4 +1,4 @@
-export * from '../../../../domain/models/exam/exam-model'
+export * from '../../../../domain/models/exam-model'
 export * from '../../../../domain/usecases/exam/delete-exam'
 export * from '../../../protocols/exam-repository/exam-repository'
 export * from '../../../protocols/exam-repository/delete-exam-repository'
