+++ src/data/usecases/exam/get-exam/db-get-exam-protocols.ts
@@ -1,4 +1,4 @@
-export * from '../../../../domain/models/exam/exam-model'
+export * from '../../../../domain/models/exam-model'
 export * from '../../../../domain/usecases/exam/get-exam'
 export * from '../../../protocols/exam-repository/exam-repository'
 export * from '../../../protocols/exam-repository/get-exam-repository'
