+++ src/data/usecases/exam/list-exams/db-list-exams-protocols.ts
@@ -1,4 +1,4 @@
-export * from '../../../../domain/models/exam/exam-model'
+export * from '../../../../domain/models/exam-model'
 export * from '../../../../domain/usecases/exam/list-exams'
 export * from '../../../protocols/exam-repository/exam-repository'
 export * from '../../../protocols/exam-repository/list-exams-repository'
