+++ src/data/usecases/exam/update-exam/db-update-exam-protocols.ts
@@ -1,3 +1,3 @@
-export * from '../../../../domain/models/exam/exam-model'
+export * from '../../../../domain/models/exam-model'
 export * from '../../../../domain/usecases/exam/update-exam'
 export * from '../../../protocols/exam-repository/update-exam-repository'
