+++ src/data/usecases/question/add-question/db-add-question-protocols.ts
@@ -1,3 +1,3 @@
-export * from '../../../../domain/models/question/question-model'
+export * from '../../../../domain/models/question-model'
 export * from '../../../../domain/usecases/question/add-question'
 export * from '../../../protocols/question-repository/add-question-repository'
