+++ src/data/usecases/question/delete-question/db-delete-question-protocols.ts
@@ -1,4 +1,4 @@
-export * from '../../../../domain/models/question/question-model'
+export * from '../../../../domain/models/question-model'
 export * from '../../../../domain/usecases/question/delete-question'
 export * from '../../../protocols/question-repository/question-repository'
 export * from '../../../protocols/question-repository/delete-question-repository'
