+++ src/data/usecases/question/get-question/db-get-question-protocols.ts
@@ -1,4 +1,4 @@
-export * from '../../../../domain/models/question/question-model'
+export * from '../../../../domain/models/question-model'
 export * from '../../../../domain/usecases/question/get-question'
 export * from '../../../protocols/question-repository/question-repository'
 export * from '../../../protocols/question-repository/get-question-repository'
