+++ src/data/usecases/question/list-questions/db-list-questions-protocols.ts
@@ -1,4 +1,4 @@
-export * from '../../../../domain/models/question/question-model'
+export * from '../../../../domain/models/question-model'
 export * from '../../../../domain/usecases/question/list-questions'
 export * from '../../../protocols/question-repository/question-repository'
 export * from '../../../protocols/question-repository/list-questions-repository'
