+++ src/data/usecases/question/update-question/db-update-question-protocols.ts
@@ -1,3 +1,3 @@
-export * from '../../../../domain/models/question/question-model'
+export * from '../../../../domain/models/question-model'
 export * from '../../../../domain/usecases/question/update-question'
 export * from '../../../protocols/question-repository/update-question-repository'
