+++ src/domain/usecases/exam/get-exam.ts
@@ -1,4 +1,4 @@
-import { ExamModel } from '../../models/exam/exam-model'
+import { ExamModel } from '../../models/exam-model'
 
 export interface GetExam {
   get: (id: string) => Promise<ExamModel>
