+++ src/domain/usecases/question/get-question.ts
@@ -1,4 +1,4 @@
-import { QuestionModel } from '../../models/question/question-model'
+import { QuestionModel } from '../../models/question-model'
 
 export interface GetQuestionModel {
   id: string
