+++ src/domain/usecases/question/list-questions.ts
@@ -1,4 +1,4 @@
-import { QuestionModel } from '../../models/question/question-model'
+import { QuestionModel } from '../../models/question-model'
 
 export interface ListQuestionsModel {
   examId: string
