+++ src/presentation/controllers/exam/exam-controller-protocols.ts
@@ -1,4 +1,4 @@
 export * from '../../protocols'
 export * from '../../protocols/exam-type-validator'
 export * from '../../../domain/usecases/exam'
-export * from '../../../domain/models/exam/exam-model'
+export * from '../../../domain/models/exam-model'
