+++ src/presentation/controllers/question/question-controller-protocols.ts
@@ -1,3 +1,3 @@
 export * from '../../protocols'
 export * from '../../../domain/usecases/question'
-export * from '../../../domain/models/question/question-model'
+export * from '../../../domain/models/question-model'
